protokollfuehrer-agent/requirements.txt
psycopg2-binary
google-generativeai
python-dotenv